refactor(AppRestorationServiceable): Moved to kDriveCore
